+++ CHANGES.md
@@ -1,4 +1,5 @@
 # 6.0.29 ????-??-??
+- 2020-05-08 Improved random number generator.
 - 2020-04-27 Fixed bug#[14997](https://bugs.otrs.org/show_bug.cgi?id=14997) - Problem with blur of RichText editor when it is left empty.
 - 2020-04-13 Fixed bug#[15035](https://bugs.otrs.org/show_bug.cgi?id=15035) - Perl Moo library missing in CheckModules script.